+++ resources/ReleaseNotes.md
@@ -2,6 +2,7 @@
 
 - Add compression flag to `mag.Science`
 - Add settings file to specify name of `timetable` properties for `mag.Science`
+- Add `Compression` property to supported events for plotting
 - Create `mag.meta.Mode` enumeration to capture sensor science mode
 - Move definition of time constants to shared utility file `mag.time.Constant`
 - Make `mag.process.Range` and `mag.process.SignedInteger` more flexible to custom variable names